+++ src/app/artwork-qc/page.tsx
@@ -1,0 +1,6 @@
+"use client";
+import ArtworkQCLayout from '@/components/ArtworkQCLayout';
+
+export default function ArtworkQC() {
+  return <ArtworkQCLayout />;
+}